fixed the problem in the collision
fixed issue with the dirbhor/dirbver and dirahor/diraver variables => when vperx == 0 and vpery1 == 0, because the velocity == 0, there' was an issue with the direction when the pokeballs switch their velocity on the perpendicular axis. This problem does not apply when just one of the variables is equal to 0, 'cause either the balls are aligned horizontally or vertically and there's no need to worry about the opposite axis.